sync patch circular import fix2
