+++ android_notify/internal/helper.py
@@ -1,7 +1,7 @@
 """Collection of useful functions"""
 
 import inspect, os, re
-from .logger import logger, on_android_platform
+from .logger import logger
 
 def can_accept_arguments(func, *args, **kwargs):
     try:
@@ -54,24 +54,3 @@ def execute_callback(callback,arg, from_who="user"):
             callback()
     except Exception as on_permissions_result_callback_error:
         logger.exception(on_permissions_result_callback_error)
-
-def on_pydroid_app():
-    package_name = "ru.iiec.pydroid3"
-    if package_name in os.environ.get("PYTHONHOME",""):
-        return True
-    elif package_name in os.path.dirname(os.path.abspath(__file__)):
-        return True
-    elif on_android_platform():
-        from android_notify.config import get_package_name
-        return package_name == get_package_name()
-    return False
-
-
-def has_androidx_dependency():
-    """Check if androidx dependencies are available"""
-    try:
-        from jnius import autoclass
-        autoclass('androidx.core.app.NotificationCompat')
-        return True
-    except Exception:
-        return False
